Moved (to WPKit) and refactored notification sync tests
