+++ WordPressKit/WordPressKitTests/Mock Data/notifications-last-seen.json
@@ -1,0 +1,4 @@
+{
+    "last_seen_time": "1477564823",
+    "success": true
+}
