+++ WordPressKit/WordPressKitTests/Mock Data/notifications-load-all.json
@@ -1,0 +1,100 @@
+{
+    "last_seen_time": "1477417665",
+    "number": 1,
+    "notes": [{
+              "id": 2674124016,
+              "type": "comment_like",
+              "read": 1,
+              "noticon": "\uf408",
+              "timestamp": "2016-10-27T10:40:23+00:00",
+              "icon": "https:\/\/0.gravatar.com\/avatar\/",
+              "url": "https:\/\/wordpress.com",
+              "subject": [{
+                          "text": "Someone liked your comment on New competitor: Google?",
+                          "ranges": [{
+                                     "type": "user",
+                                     "indices": [0, 11],
+                                     "url": "http:\/\/bostonsportsmedia.com",
+                                     "site_id": 107738687,
+                                     "email": "someone@blog",
+                                     "id": 47232
+                                     }, {
+                                     "type": "post",
+                                     "indices": [34, 57],
+                                     "url": "https:\/\/wordpress.com",
+                                     "site_id": 2647731,
+                                     "id": 43671
+                                     }]
+                          }],
+              "body": [{
+                       "text": "Bruce Allen",
+                       "ranges": [{
+                                  "email": "someone@blog",
+                                  "url": "http:\/\/tada.blog",
+                                  "id": 47232,
+                                  "site_id": 107738687,
+                                  "type": "user",
+                                  "indices": [0, 11]
+                                  }],
+                       "media": [{
+                                 "type": "image",
+                                 "indices": [0, 0],
+                                 "height": "256",
+                                 "width": "256",
+                                 "url": "https:\/\/0.gravatar.com\/avatar\/"
+                                 }],
+                       "actions": {
+                       "follow": false
+                       },
+                       "meta": {
+                       "links": {
+                       "email": "someone@blog",
+                       "home": "http:\/\/someone.blog"
+                       },
+                       "ids": {
+                       "user": 47232,
+                       "site": 107738687
+                       },
+                       "titles": {
+                       "home": "Boston Sports Media Watch",
+                       "tagline": "Watching the New England Sports Media Since 2002"
+                       }
+                       },
+                       "type": "user"
+                       }],
+              "meta": {
+              "ids": {
+              "site": 2647731,
+              "post": 43671,
+              "comment": 84396
+              },
+              "links": {
+              "site": "https:\/\/public-api.wordpress.com\/rest\/v1\/sites\/2647731",
+              "post": "https:\/\/public-api.wordpress.com\/rest\/v1\/posts\/43671",
+              "comment": "https:\/\/public-api.wordpress.com\/rest\/v1\/comments\/84396"
+              }
+              },
+              "title": "1 Like",
+              "header": [{
+                         "text": "Jorge Leandro Perez",
+                         "ranges": [{
+                                    "type": "user",
+                                    "indices": [0, 19],
+                                    "url": "http:\/\/www.lantean.co",
+                                    "site_id": 46233804,
+                                    "email": "someone@somewhere",
+                                    "id": 45413599
+                                    }],
+                         "media": [{
+                                   "type": "image",
+                                   "indices": [0, 0],
+                                   "height": "256",
+                                   "width": "256",
+                                   "url": "https:\/\/2.gravatar.com\/avatar\/e44d9c00c3af8037f203ca42207a99d7?s=256&d=https%3A%2F%2Fsecure.gravatar.com%2Favatar%2Fad516503a11cd5ca435acc9bb6523536%3Fs%3D256&r=G"
+                                   }]
+                         }, {
+                         "text": "Some text here please!"
+                         }],
+              "note_hash": 4007447833
+              }]
+}
