+++ WordPressKit/WordPressKitTests/Mock Data/notifications-mark-as-read.json
@@ -1,0 +1,4 @@
+{
+    "updated": [],
+    "success": true
+}
